+++ tap_shopify/__init__.py
@@ -19,6 +19,7 @@ from tap_shopify.streams.base import shopify_error_handling, get_request_timeout
 REQUIRED_CONFIG_KEYS = ["shop", "api_key"]
 LOGGER = singer.get_logger()
 SDC_KEYS = {'id': 'integer', 'name': 'string', 'myshopify_domain': 'string'}
+UNSUPPORTED_FIELDS = {"author"}
 
 @shopify_error_handling
 def initialize_shopify_client():
@@ -33,6 +34,30 @@ def initialize_shopify_client():
 
     # Shop.current() makes a call for shop details with provided shop and api_key
     return shopify.Shop.current().attributes
+
+# Add helper
+def fetch_app_scopes():
+    query = """
+    query {
+      currentAppInstallation {
+        accessScopes {
+          handle
+        }
+      }
+    }
+    """
+    data = json.loads(shopify.GraphQL().execute(query))
+    return {s["handle"] for s in data["data"]["currentAppInstallation"]["accessScopes"]}
+
+def has_read_users_access():
+    # If the app does not have the 'read_users' scope, return False
+    if 'read_users' not in fetch_app_scopes():
+        LOGGER.warning(
+            "Skipping '%s' field: 'read_users' scope is not granted for public apps.",
+            ", ".join(UNSUPPORTED_FIELDS)
+        )
+        return False
+    return True
 
 def get_abs_path(path):
     return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
@@ -63,6 +88,8 @@ def get_discovery_metadata(stream, schema):
     for field_name in schema['properties'].keys():
         if field_name in stream.key_properties or field_name == stream.replication_key:
             mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
+        elif field_name in UNSUPPORTED_FIELDS and not has_read_users_access():
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'unsupported')
         else:
             mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'available')